AI: some methods. AIRunner: new constructor that takes AI instance. Event: getTime() now returns long. Game: mainLoop implementation. GameQueue: fixed to fit new Event contract. GraphicsEngine: added init() method. Player: now manages a AIRunner. Skill: fixed to fit new Event contract. Updater: many implementations.

+++ PA/src/Event.java
@@ -6,7 +6,7 @@
  */
 public interface Event {
 	/** The time in millisecond after game starts that this event should happen */
-	int getTime();
+	long getTime();
 	
 	/** What should the event do */
 	void action();
